fix: move the navigate function to a dedicate file only for frontend use
