+++ plugins/woocommerce-blocks/assets/js/blocks/collection-filters/frontend.ts
@@ -1,0 +1,15 @@
+/**
+ * External dependencies
+ */
+import { navigate as navigateFn } from '@woocommerce/interactivity';
+import { getSetting } from '@woocommerce/settings';
+
+const isBlockTheme = getSetting< boolean >( 'isBlockTheme' );
+const isProductArchive = getSetting< boolean >( 'isProductArchive' );
+
+export function navigate( href: string, options = {} ) {
+	if ( ! isBlockTheme && isProductArchive ) {
+		return ( window.location.href = href );
+	}
+	return navigateFn( href, options );
+}
